\name{gbp-internal}
\alias{alpha.est.prior.kn}
\alias{alpha.est.prior.un}
\alias{bp}
\alias{br.ini.kn}
\alias{br.ini.un}
\alias{br.llik.prior.kn}
\alias{br.llik.prior.un}
\alias{br.post.est.prior.kn}
\alias{br.post.est.prior.un}
\alias{gr.cp.to.dp}
\alias{gr.ll.muknown}
\alias{gr.ll.muunknown}
\alias{gr}
\alias{pr.ini.kn}
\alias{pr.ini.un}
\alias{pr.llik.prior.kn}
\alias{pr.llik.prior.un}
\alias{pr.llik.prior.un}
\alias{shrink.est}
\alias{pr.llik.prior.un}

\title{Internal GBP functions}
\description{Internal GBP functions}
\Usage{}
\details{These are not to be called by the user.}
\keyword{internal}

