\name{gbp-internal}
\alias{BRInitialValue2ndLevelMeanKnown}
\alias{BRInitialValue2ndLevelMeanUnknown}
\alias{BRAlphaEst2ndLevelMeanKnown}
\alias{BRAlphaBetaEst2ndLevelMeanUnknown}
\alias{BRShrinkageEst}
\alias{BRPosteriorEst2ndLevelMeanKnown}
\alias{BRPosteriorEst2ndLevelMeanUnknown}
\alias{BRIS}
\alias{BRIS2ndLevelMeanKnown}
\alias{br}

\alias{PRInitialValue2ndLevelMeanKnown}
\alias{PRInitialValue2ndLevelMeanUnknown}
\alias{PRAlphaEst2ndLevelMeanKnown}
\alias{PRAlphaBetaEst2ndLevelMeanUnknown}
\alias{PRShrinkageEst}
\alias{PRPosteriorEst2ndLevelMeanKnown}
\alias{PRPosteriorEst2ndLevelMeanUnknown}
\alias{pr}

\alias{gr.ll.muknown}
\alias{gr.ll.muunknown}
\alias{gr.cp.to.dp}
\alias{derval}
\alias{tr}
\alias{gr}

\title{Internal \code{gbp} functions}
\description{Internal \code{gbp} functions}
\details{These are not to be called by the user.}
\keyword{internal}

