\name{plot.gbp}
\alias{plot.gbp}
\title{
Shrinkage and Posterior Interval Plots
}
\description{
This function plots two graphs, Shrinkage and Posterior Interval Plots, based on the object defined as "gbp" class.
}
\usage{
\method{plot}{gbp}(x,...)
}
\arguments{
  \item{x}{
	The result of gbp function, which has "gbp" class
}
  \item{...}{
	Any argument for future purpose
}
}
\value{
If an object defined as "gbp" class is put into this function, then two plots, shrinkage and posterior interval plots, will show up.
}
\author{
Joseph Kelly and Hyungsuk Tak
}
\keyword{methods}
