\name{schools}
\alias{schools}
\docType{data}

\title{
Eight Schools Data
}

\description{
Dataset as seen in Rubin (1981) which was an analysis of coaching
effects on SAT scores from eight schools.
}

\usage{data(schools)}

\format{
  A dataset of 8 schools containing
  \describe{
    \item{\code{y}}{The observed coaching effect of each school}
    \item{\code{se}}{The standard error of the coaching effect of each school.}
  }
}

\source{
Rubin, D. B. (1981). \emph{Estimation in parallel randomized
  experiments.} Journal of Educational Statistics, 6:377-401.
}
\references{
Rubin, D. B. (1981). \emph{Estimation in parallel randomized
  experiments.} Journal of Educational Statistics, 6:377-401.
}
\examples{
data(schools)
}

\keyword{datasets}